docker/airflow/requirements.txt
konlpy
fasttext-wheel
beautifulsoup4==4.12.2
lxml             # HTML 파싱 성능 향상을 위해 (선택)
pandas           # 데이터 처리용 (필요하면)
optuna
scikit-learn
transformers[torch] 
scikit-learn
fastapi==0.110.0
uvicorn==0.29.0
fugashi[unidic-lite]
accelerate

mysql-connector-python==8.0.33
requests==2.31.0
psutil==5.9.0
python-dateutil==2.8.2
pytz==2023.3